LPS-114698 Renames base classes
