+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/data/set/view/list/BaseListClayDataSetDisplayView.java
@@ -12,38 +12,32 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.data.set.view.table;
+package com.liferay.frontend.taglib.clay.data.set.view.list;
 
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
 import com.liferay.frontend.taglib.clay.data.set.constants.ClayDataSetConstants;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
-
-import java.util.Locale;
-import java.util.ResourceBundle;
+import com.liferay.petra.string.StringPool;
 
 /**
- * @author Marco Leo
+ * @author Alessio Antonio Rendina
  */
-public abstract class ClayTableDataSetDisplayView
+public abstract class BaseListClayDataSetDisplayView
 	implements ClayDataSetDisplayView {
 
-	public abstract ClayTableSchema getClayTableSchema();
-
 	public String getContentRenderer() {
-		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
+		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_LIST;
 	}
+
+	public abstract String getDescription();
 
 	public String getLabel() {
-		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
-	}
-
-	public ResourceBundle getResourceBundle(Locale locale) {
-		return ResourceBundleUtil.getBundle(
-			"content.Language", locale, getClass());
+		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_LIST;
 	}
 
 	public String getThumbnail() {
-		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
+		return StringPool.BLANK;
 	}
+
+	public abstract String getTitle();
 
 }
