+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/java/com/liferay/frontend/taglib/clay/data/set/view/table/BaseTableClayDataSetDisplayView.java
@@ -12,36 +12,38 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.clay.data.set.view.table.selectable;
+package com.liferay.frontend.taglib.clay.data.set.view.table;
 
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
 import com.liferay.frontend.taglib.clay.data.set.constants.ClayDataSetConstants;
-import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
+import java.util.ResourceBundle;
 
 /**
- * @author Alessio Antonio Rendina
+ * @author Marco Leo
  */
-public abstract class ClaySelectableTableDataSetDisplayView
+public abstract class BaseTableClayDataSetDisplayView
 	implements ClayDataSetDisplayView {
 
+	public abstract ClayTableSchema getClayTableSchema();
+
 	public String getContentRenderer() {
-		return ClayDataSetConstants.
-			CLAY_DATA_SET_CONTENT_RENDERER_SELECTABLE_TABLE;
+		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
 	}
 
-	public abstract String getFirstColumnLabel(Locale locale);
-
-	public abstract String getFirstColumnName();
-
 	public String getLabel() {
-		return ClayDataSetConstants.
-			CLAY_DATA_SET_CONTENT_RENDERER_SELECTABLE_TABLE;
+		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
+	}
+
+	public ResourceBundle getResourceBundle(Locale locale) {
+		return ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
 	}
 
 	public String getThumbnail() {
-		return StringPool.BLANK;
+		return ClayDataSetConstants.CLAY_DATA_SET_CONTENT_RENDERER_TABLE;
 	}
 
 }
